third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-allow-in-http-header-expected.txt

CONSOLE ERROR: Unrecognized Content-Security-Policy directive 'suborigin'.

CONSOLE ERROR: Unrecognized Content-Security-Policy directive 'suborigin'.

CONSOLE ERROR: Unrecognized Content-Security-Policy directive 'suborigin'.


This is a testharness.js-based test.
FAIL Allow suborigin in HTTP header Uncaught assert_unreached: Reached unreachable code
Harness: the test ran to completion.
